Move index & helper functions from SleepTracker to Tracker file #626 #686
